requirements.txt
flask>=3.0
pillow>=10.0
numpy>=1.26
tensorflow>=2.15  
opencv-python>=4.10